requirements.txt
asgiref==3.8.1
Django==4.2.15
sqlparse==0.5.1
typing-extensions==4.12.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
